rename evaluate script to test
